Revert "memory optimizations"
This reverts commit ca3a1171a809beb1b4e1edfe8fbb2eca667f8ee6.

Conflicts:
	RSSIAntennaTracker.ino

+++ RSSIAntennaTracker.ino
@@ -1,7 +1,7 @@
 #include "Config.h"
 
-uint8_t servoPositionPan = SERVO_MID;
-uint8_t servoPositionTilt = SERVO_MID;
+int servoPositionPan = SERVO_MID;
+int servoPositionTilt = SERVO_MID;
 
 void setup()
 {
